:recycle: refactor: abstract custom tooltip into component
